feat: moved swap handler to ptx handler lld
