+++ .changeset/silver-feet-flow.md
@@ -1,0 +1,5 @@
+---
+"ledger-live-desktop": minor
+---
+
+feat: moved swap handler to PTX handler
